Split lowering/lower.rs into lower/exp.rs and lower/decls.rs
